src/requirements/common.txt
# core
gunicorn==19.4.5
raven==5.11.2
requests==2.9.1
Django==1.8.11
psycopg2>=2.6
kitchen==1.2.4
django-mptt==0.8.3
django-mptt-admin==0.3.4

# admin
django-grappelli==2.8.1

# django-rest-framework
djangorestframework==3.3.3
djangorestframework-oauth==1.1.0
django-filter==0.13.0
drf-extensions

# cache search queues
redis==2.10.5
django-redis==4.3.0
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
click==6.3
rq==0.5.6
django_rq==0.9.0
django-widget-tweaks

# AWS imaging
Pillow==2.8.1
boto==2.39.0

# google
httplib2==0.9.2
google-api-python-client==1.5.0

# push + sockets
pusher==1.2.3
django-cors-headers==1.1.0
django-pgjson==0.3.1

# tracking
mixpanel==4.3.0

# docs
django-rest-swagger==0.3.4
markdown==2.6.6

# utils
django-request-id
pydash
phonenumbers==7.2.8
django-ipware
twilio==6.2.dev0
nexmo

# sss
# requires apt-get install libcurl4-openssl-dev libxml2-dev libxslt-dev
antigate
pyquery
fake_useragent

# urllib
# requires apt-get install libffi-dev
pyopenssl
ndg-httpsclient
pyasn1

# python interface for django-admin shell (v 0.15 breaks)
bpython==0.14.2

# http://stackoverflow.com/questions/34646942/ssl3-get-server-certificate-certificate-verify-failed-on-python-when-requesting
certifi==2015.04.28
